+++ src/main/java/cat/soft/src/parking/model/room/GetJoinRoomRes.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 @Getter
 @Setter
 public class GetJoinRoomRes {
-	private Integer roomIdx;
+	private Long roomIdx;
 }
